2019.06.14 (1.52p52; Multi-point selections)

+++ ij/plugin/Selection.java
@@ -32,11 +32,12 @@ public class Selection implements PlugIn, Measurements {
 		if (imp==null)
 			{IJ.noImage(); return;}
 		if (arg.equals("all")) {
-			imp.saveRoi();
-			imp.setRoi(0,0,imp.getWidth(),imp.getHeight());
-		} else if (arg.equals("none"))
-			imp.deleteRoi();
-		else if (arg.equals("restore"))
+			if (imp.okToDeleteRoi())
+				imp.setRoi(0,0,imp.getWidth(),imp.getHeight());
+		} else if (arg.equals("none")) {
+			if (imp.okToDeleteRoi())
+				imp.deleteRoi();
+		} else if (arg.equals("restore"))
 			imp.restoreRoi();
 		else if (arg.equals("spline"))
 			fitSpline();
@@ -71,6 +72,7 @@ public class Selection implements PlugIn, Measurements {
 	}
 	
 	private void rotate(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		Roi roi = imp.getRoi();
 		if (IJ.macroRunning()) {
 			String options = Macro.getOptions();
@@ -83,6 +85,7 @@ public class Selection implements PlugIn, Measurements {
 	}
 	
 	private void enlarge(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		Roi roi = imp.getRoi();
 		if (roi!=null) {
 			Undo.setup(Undo.ROI, imp);
@@ -102,6 +105,7 @@ public class Selection implements PlugIn, Measurements {
 	Authors: Nikolai Chernov, Michael Doube, Ved Sharma
 	*/
 	void fitCircle(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		Roi roi = imp.getRoi();
 		if (roi==null) {
 			noRoi("Fit Circle");
@@ -455,6 +459,7 @@ public class Selection implements PlugIn, Measurements {
 	}
 
 	void createEllipse(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		IJ.showStatus("Fitting ellipse");
 		Roi roi = imp.getRoi();
 		if (roi==null)
@@ -480,6 +485,7 @@ public class Selection implements PlugIn, Measurements {
 	}
 
 	void convexHull(ImagePlus imp) {
+		if (!imp.okToDeleteRoi()) return;
 		long startTime = System.currentTimeMillis();
 		Roi roi = imp.getRoi();
 		if (roi == null) { IJ.error("Convex Hull", "Selection required"); return; }
@@ -693,6 +699,8 @@ public class Selection implements PlugIn, Measurements {
 			noRoi("To Bounding Box");
 			return;
 		}
+		if (!imp.okToDeleteRoi())
+			return;
 		Undo.setup(Undo.ROI, imp);
 		Rectangle r = roi.getBounds();
 		imp.deleteRoi();
